requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
chromadb
streamlit
requests
python-dotenv